msilo: cast to fix storage size warning
(cherry picked from commit c11fcbbe9ed19bc54b800100d15d9ae269735f0c)

+++ src/modules/msilo/msilo.c
@@ -755,7 +755,7 @@ static int m_store(sip_msg_t* msg, str *owner_s)
 	}
 
 	/* current time */
-	val = (int)time(NULL);
+	val = (int)(unsigned long long)time(NULL);
 
 	/* add expiration time */
 	db_keys[nr_keys] = &sc_exp_time;
